Factor boolean parsing into separate package-private class
